+++ src/evm/protocol/uniswap_v2/state.rs
@@ -1,7 +1,8 @@
 use std::{any::Any, collections::HashMap};
 
-use alloy_primitives::U256;
+use alloy_primitives::{Address, U256};
 use num_bigint::{BigUint, ToBigUint};
+use num_traits::Zero;
 use tycho_core::{dto::ProtocolStateDelta, Bytes};
 
 use super::reserve_price::spot_price_from_reserves;
@@ -98,6 +99,46 @@ impl ProtocolSim for UniswapV2State {
                 .expect("Expected an unsigned integer as gas value"),
             Box::new(new_state),
         ))
+    }
+
+    fn get_limits(
+        &self,
+        token_in: Address,
+        token_out: Address,
+    ) -> Result<(BigUint, BigUint), SimulationError> {
+        if self.reserve0 == U256::from(0u64) || self.reserve1 == U256::from(0u64) {
+            return Ok((BigUint::zero(), BigUint::zero()));
+        }
+
+        let zero_for_one = token_in < token_out;
+        let (reserve_in, reserve_out) = if zero_for_one {
+            (self.reserve0, self.reserve1)
+        } else {
+            (self.reserve1, self.reserve0)
+        };
+
+        // Soft limit for amount in is the amount to get a 90% price impact.
+        // The two equations to resolve are:
+        // - 90% price impact: (reserve1 - y)/(reserve0 + x) = 0.1 × (reserve1/reserve0)
+        // - Maintain constant product: (reserve0 + x) × (reserve1 - y) = reserve0 * reserve1
+        //
+        // This resolves into x = (√10 - 1) × reserve0 = 2.16 × reserve0
+        let amount_in =
+            safe_div_u256(safe_mul_u256(reserve_in, U256::from(216))?, U256::from(100))?;
+
+        // Calculate amount_out using the constant product formula
+        // The constant product formula requires:
+        // (reserve_in + amount_in) × (reserve_out - amount_out) = reserve_in * reserve_out
+        // Solving for amount_out:
+        // amount_out = reserve_out - (reserve_in * reserve_out) (reserve_in + amount_in)
+        // which simplifies to:
+        // amount_out = (reserve_out * amount_in) / (reserve_in + amount_in)
+        let amount_out = safe_div_u256(
+            safe_mul_u256(reserve_out, amount_in)?,
+            safe_add_u256(reserve_in, amount_in)?,
+        )?;
+
+        Ok((u256_to_biguint(amount_in), u256_to_biguint(amount_out)))
     }
 
     fn delta_transition(
@@ -332,5 +373,51 @@ mod tests {
             }
             _ => panic!("Test failed: was expecting an Err value"),
         };
+    }
+
+    #[test]
+    fn test_get_limits_price_impact() {
+        let state =
+            UniswapV2State::new(U256::from_str("1000").unwrap(), U256::from_str("100000").unwrap());
+
+        let (amount_in, _) = state
+            .get_limits(
+                Address::from_str("0x0000000000000000000000000000000000000000").unwrap(),
+                Address::from_str("0x0000000000000000000000000000000000000001").unwrap(),
+            )
+            .unwrap();
+
+        let token_0 = Token::new(
+            "0x0000000000000000000000000000000000000000",
+            18,
+            "T0",
+            10_000.to_biguint().unwrap(),
+        );
+        let token_1 = Token::new(
+            "0x0000000000000000000000000000000000000001",
+            18,
+            "T1",
+            10_000.to_biguint().unwrap(),
+        );
+
+        let result = state
+            .get_amount_out(amount_in.clone(), &token_0, &token_1)
+            .unwrap();
+        let new_state = result
+            .new_state
+            .as_any()
+            .downcast_ref::<UniswapV2State>()
+            .unwrap();
+
+        let initial_price = state
+            .spot_price(&token_0, &token_1)
+            .unwrap();
+        let new_price = new_state
+            .spot_price(&token_0, &token_1)
+            .unwrap()
+            .floor();
+
+        let expected_price = initial_price / 10.0;
+        assert!(expected_price == new_price, "Price impact not 90%.");
     }
 }